requirements.txt
requests
googlesearch-python
google-genai
python-dotenv
streamlit